+++ src/test_fixtures/mod.rs
@@ -17,6 +17,8 @@ pub mod events;
 #[allow(clippy::must_use_candidate)]
 pub mod logging;
 #[allow(clippy::must_use_candidate)]
+pub mod manifest;
+#[allow(clippy::must_use_candidate)]
 pub mod mocks;
 #[allow(clippy::must_use_candidate)]
 pub mod mtls;